Patch 1.0.1
- WebGL build | P to pause
- Changed unreadable font in Dialogs | Improved font for StepsMenu
- Fixed Options in PauseMenu

+++ Chroma Bounce/Assets/_Project/Scripts/Core/SaveSerial.cs
@@ -31,14 +31,18 @@ public class SaveSerial : MonoBehaviour{	public static SaveSerial instance;
 	public string _playerDataPath(){return Application.persistentDataPath+"/"+filename+".hyper";}
 	public void Save(){
         var settings=new ES3Settings(_playerDataPath(),ES3.EncryptionType.None);
+		//#if UNITY_WEBGL || UNITY_WEBGL_API
+		if(Application.platform==RuntimePlatform.WebGLPlayer){settings=new ES3Settings(ES3.Location.PlayerPrefs);}
+		//#endif
 		if(!ES3.KeyExists("buildFirstLoaded",settings)){buildFirstLoaded=GameManager.instance.buildVersion;ES3.Save("buildFirstLoaded",buildFirstLoaded,settings);}
 		buildLastLoaded=GameManager.instance.buildVersion;ES3.Save("buildLastLoaded",buildLastLoaded,settings);
 		ES3.Save("playerData",playerData,settings);
 		Debug.Log("Game Data saved");
 	}
 	public void Load(){
-		if(ES3.FileExists(_playerDataPath())){
+		if(ES3.FileExists(_playerDataPath())||Application.platform==RuntimePlatform.WebGLPlayer){
 			var settings=new ES3Settings(_playerDataPath(),ES3.EncryptionType.None);
+			if(Application.platform==RuntimePlatform.WebGLPlayer){settings=new ES3Settings(ES3.Location.PlayerPrefs);}
 
 			if(ES3.KeyExists("buildFirstLoaded",settings)){buildFirstLoaded=ES3.Load<float>("buildFirstLoaded",8,settings);}
 			else{Debug.LogWarning("Key for buildFirstLoaded not found in: "+_playerDataPath());}
@@ -57,7 +61,7 @@ public class SaveSerial : MonoBehaviour{	public static SaveSerial instance;
 		RecreatePlayerData();
 		Debug.Log("Game Data reset");
 		GC.Collect();
-		if(ES3.FileExists(_playerDataPath())){
+		if(ES3.FileExists(_playerDataPath())&&Application.platform!=RuntimePlatform.WebGLPlayer){
 			ES3.DeleteFile(_playerDataPath());
 			Debug.Log("Game Data deleted!");
 		}
@@ -102,12 +106,14 @@ public class SaveSerial : MonoBehaviour{	public static SaveSerial instance;
 	public string _settingsDataPath(){return Application.persistentDataPath+"/"+filenameSettings+".json";}
 	public void SaveSettings(){
 		var settings=new ES3Settings(_settingsDataPath(),ES3.EncryptionType.None);
+		if(Application.platform==RuntimePlatform.WebGLPlayer){settings=new ES3Settings(ES3.Location.PlayerPrefs);}
 		ES3.Save("settingsData",settingsData,settings);
 		Debug.Log("Settings saved");
 	}
 	public void LoadSettings(){
-		if(ES3.FileExists(_settingsDataPath())){
+		if(ES3.FileExists(_settingsDataPath())||Application.platform==RuntimePlatform.WebGLPlayer){
 		var settings=new ES3Settings(_settingsDataPath(),ES3.EncryptionType.None);
+		if(Application.platform==RuntimePlatform.WebGLPlayer){settings=new ES3Settings(ES3.Location.PlayerPrefs);}
 			if(ES3.KeyExists("settingsData",settings)){ES3.LoadInto<SettingsData>("settingsData",settingsData,settings);}
 			else{Debug.LogWarning("Key for settingsData not found in: "+_settingsDataPath());}
 		}else Debug.LogWarning("Settings file not found in: "+_settingsDataPath());
@@ -115,7 +121,7 @@ public class SaveSerial : MonoBehaviour{	public static SaveSerial instance;
 	public void DeleteSettings(){
 		settingsData=new SettingsData();
 		GC.Collect();
-		if(ES3.FileExists(_settingsDataPath())){
+		if(ES3.FileExists(_settingsDataPath())&&Application.platform!=RuntimePlatform.WebGLPlayer){
 			ES3.DeleteFile(_settingsDataPath());
 			Debug.Log("Settings deleted");
 		}
